+++ platformio_nuli_avionics_flight_software/src/drivers/arduino/SerialDebug.h
@@ -22,7 +22,7 @@ public:
         Serial.print(str);
     }
 
-    void print(int64_t num) override {
+    void print(int32_t num) override {
         Serial.print(num);
     }
